+++ app/src/main/java/com/antisnusbolaget/slutasnusa2/di/AppModule.kt
@@ -2,7 +2,7 @@ package com.antisnusbolaget.slutasnusa2.di
 
 import com.antisnusbolaget.slutasnusa2.viewmodel.AchievementViewModel
 import com.antisnusbolaget.slutasnusa2.viewmodel.HomeScreenViewModel
-import com.antisnusbolaget.slutasnusa2.viewmodel.OnBoardingViewModel
+import com.antisnusbolaget.slutasnusa2.viewmodel.onboarding.OnBoardingViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
