+++ app/src/main/java/com/antisnusbolaget/slutasnusa2/ui/screens/mainscreen/homescreen/HomeScreen.kt
@@ -4,26 +4,32 @@ import android.content.res.Configuration
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Devices
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.sp
+import com.antisnusbolaget.slutasnusa2.ui.components.TextBold
+import com.antisnusbolaget.slutasnusa2.ui.theme.yellow
 
 @Composable
-fun AchievementScreen() {
+fun HomeScreen() {
     Box(
         modifier = Modifier
-            .fillMaxSize()
-            .background(Color.Gray),
+            .fillMaxSize(),
     ) {
-        Text(text = "AchievementScreen", fontSize = 30.sp)
+        Box(
+            contentAlignment = Alignment.Center,
+            modifier = Modifier
+                .fillMaxSize()
+                .background(yellow),
+        ) {
+            TextBold(text = "HOME SCREEN CONTENT")
+        }
     }
 }
 
-private const val componentName = "AchievementScreen"
+private const val componentName = "Home Screen"
 
 @Preview("$componentName (light)", showBackground = true)
 @Preview("$componentName (dark)", uiMode = Configuration.UI_MODE_NIGHT_YES)
@@ -31,5 +37,5 @@ private const val componentName = "AchievementScreen"
 @Preview("$componentName (large screen)", device = Devices.PIXEL_C)
 @Composable
 private fun PreviewComponent() {
-    AchievementScreen()
+    HomeScreen()
 }
